# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiBoleta API",
        "description": "API para gestión de documentos y vacaciones",
        "contact": {
            "email": "admin@miboleta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/batches": {
            "get": {
                "tags": [
                    "Lotes de Documentos"
                ],
                "summary": "Listar lotes de carga",
                "description": "Lista el historial de cargas masivas de documentos",
                "operationId": "a19a3f194ffbb23d02e34e8beb90d824",
                "parameters": [
                    {
                        "name": "X-Tenant-Id",
                        "in": "header",
                        "description": "ID del tenant",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "completed",
                                "failed"
                            ]
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de documento",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lotes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DocumentBatch"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado (solo admin/root)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/batches/{id}": {
            "get": {
                "tags": [
                    "Lotes de Documentos"
                ],
                "summary": "Obtener detalle de lote",
                "description": "Obtiene los detalles de un lote incluyendo errores y estadísticas",
                "operationId": "6c8c7e5e8a9e29e82da87832e7e8008e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lote",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del lote",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentBatch"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Lote no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/batches/upload": {
            "post": {
                "tags": [
                    "Lotes de Documentos"
                ],
                "summary": "Subir lote de documentos",
                "description": "Sube un archivo ZIP con documentos PDF para procesamiento asíncrono. Los nombres de archivos deben ser el número de documento del empleado (ej: 12345678.pdf)",
                "operationId": "9c834cd673a7bc44be20168f7d708530",
                "parameters": [
                    {
                        "name": "X-Tenant-Id",
                        "in": "header",
                        "description": "ID del tenant",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "type_id",
                                    "period"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo ZIP con PDFs",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type_id": {
                                        "description": "ID del tipo de documento",
                                        "type": "integer"
                                    },
                                    "period": {
                                        "description": "Periodo (ej: 2024-01)",
                                        "type": "string"
                                    },
                                    "notify_employees": {
                                        "description": "Notificar a empleados por email",
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "requires_signature": {
                                        "description": "Requiere firma digital",
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Archivo recibido para procesamiento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Archivo recibido. El procesamiento comenzará en breve."
                                        },
                                        "data": {
                                            "properties": {
                                                "batch_id": {
                                                    "type": "integer"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Tenant no especificado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/batches/preview": {
            "post": {
                "tags": [
                    "Lotes de Documentos"
                ],
                "summary": "Vista previa de ZIP",
                "description": "Analiza el contenido de un archivo ZIP antes de procesar, mostrando archivos válidos e inválidos",
                "operationId": "03f03fc7d77f04d89923215f5051cf58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo ZIP a analizar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Análisis del ZIP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_files": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "valid_pdfs": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "invalid_names": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "invalid_formats": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se pudo abrir el archivo ZIP"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentBatch": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tenant_id": {
                        "type": "string"
                    },
                    "uploaded_by": {
                        "type": "string"
                    },
                    "doc_type_id": {
                        "type": "integer"
                    },
                    "document_type": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "display_name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "period": {
                        "type": "string",
                        "example": "2024-01"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "file_path": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ]
                    },
                    "total_files": {
                        "type": "integer"
                    },
                    "processed_files": {
                        "type": "integer"
                    },
                    "failed_files": {
                        "type": "integer"
                    },
                    "orphan_files": {
                        "type": "integer"
                    },
                    "progress": {
                        "description": "Porcentaje de progreso 0-100",
                        "type": "integer"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "processed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}